+++ api/src/main/java/io/strimzi/api/kafka/model/podset/StrimziPodSet.java
@@ -38,8 +38,8 @@ import java.util.Map;
                 group = StrimziPodSet.RESOURCE_GROUP,
                 scope = StrimziPodSet.SCOPE,
                 versions = {
-                    @Crd.Spec.Version(name = Constants.V1, served = true, storage = false),
-                    @Crd.Spec.Version(name = Constants.V1BETA2, served = true, storage = true, deprecated = true, deprecationWarning = "Version v1beta2 of the StrimziPodSet API is deprecated. Please use the v1 version instead.")
+                    @Crd.Spec.Version(name = Constants.V1, served = true, storage = true),
+                    @Crd.Spec.Version(name = Constants.V1BETA2, served = true, storage = false, deprecated = true, deprecationWarning = "Version v1beta2 of the StrimziPodSet API is deprecated. Please use the v1 version instead.")
                 },
                 subresources = @Crd.Spec.Subresources(
                     status = @Crd.Spec.Subresources.Status()
@@ -82,7 +82,7 @@ import java.util.Map;
 @DescriptionFile
 @EqualsAndHashCode(callSuper = true)
 @ToString(callSuper = true)
-@Version(Constants.V1BETA2)
+@Version(Constants.V1)
 @Group(Constants.RESOURCE_CORE_GROUP_NAME)
 public class StrimziPodSet extends CustomResource<StrimziPodSetSpec, StrimziPodSetStatus> implements Namespaced, UnknownPropertyPreserving {
     private static final long serialVersionUID = 1L;